use gtest fused source

+++ test/lv5/test_radio_arena.cc
@@ -7,6 +7,8 @@ TEST(RadioArenaCase, ArenaBlocksTest) {
   // check blocks size and iterator / const_iterator behavior
   iv::core::ScopedPtr<iv::lv5::radio::Arena> arena(new iv::lv5::radio::Arena(NULL));
 
-  EXPECT_EQ(iv::lv5::radio::kBlocks, std::distance(arena->begin(), arena->end()));
-  EXPECT_EQ(iv::lv5::radio::kBlocks, std::distance(arena->cbegin(), arena->cend()));
+  EXPECT_EQ(iv::lv5::radio::kBlocks,
+            static_cast<std::size_t>(std::distance(arena->begin(), arena->end())));
+  EXPECT_EQ(iv::lv5::radio::kBlocks,
+            static_cast<std::size_t>(std::distance(arena->cbegin(), arena->cend())));
 }
